+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All @usebones packages release together with a single version (see
 
 ## Unreleased
 
+- icons: add arrow-up and arrow-down icons.
+
 ## 0.2.1 (2026-08-26)
 
 - tokens, react, icons: the library name is capitalized (Bones) across